@@ -1,9 +1,9 @@
 package ruler
 
 import (
-	"github.com/cortexproject/cortex/pkg/ruler"
-	"github.com/cortexproject/cortex/pkg/ruler/rulestore"
 	"github.com/go-kit/kit/log"
+	"github.com/grafana/dskit/ruler"
+	"github.com/grafana/dskit/ruler/rulestore"
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/grafana/loki/pkg/logql"
